Added support for inheritcance from a single REST endpoint
Look for the new NSEntityDescription UserInfo key jsonUseAbstract and the new FOSRESTServiceAdapter method subtypeFromBase:givenJSON:.  These two methods work together to disambiguate resolving an abstract entity to an concrete one based upon the JSON data received from the REST service.

Also fixed a bug in the fast tracking code that would sometimes mark an object as fast-tracked when it really was not.