Added configuration support for whether offline change are saved
Previously offline changes were discarded as synchronization takes place from the user point down.  So, if there's no logged in user, then we cannot synchronize the changes.  However, it's possible that the user might want to populate some tables before the user logs in (static data, for example).  Normally static tables are loaded when the user logs in, however if some table data is needed from the server prior to the user logging in, then this option can be turned on to allow those tables to be populated.