Fixed a bug with the newly introduced feature of saving offline changes
While we want to enable the saving of changes while offline, we don't want to do it while we're logging out as the local updates are irrelevant (e.g. just state updates received as responses from the server during the logout process).  If we don't ignore these changes, then, if we're deleting the local cache after logging out there will be changes left in the MOC when the store coordinator is changed leaving dangling data in the MOC.